# Improve performance of metrics hashing in statsd module