--- a/metricbeat/helper/labelhash/hash.go
+++ b/metricbeat/helper/labelhash/hash.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package prometheus
+package labelhash
 
 import (
 	"bytes"