--- a/metricbeat/helper/labelhash/hash_benchmark_test.go
+++ b/metricbeat/helper/labelhash/hash_benchmark_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package prometheus
+package labelhash
 
 import (
 	"testing"