## What does this PR do?
This moves a bit of hashing code used for prometheus labels
around and changes statsd metric tag hashing from being plain json
to using the xxhash for improved performance.


```
name      old time/op    new time/op    delta
Ingest-4    3.45µs ± 1%    2.81µs ±13%  -18.57%  (p=0.008 n=5+5)

name      old alloc/op   new alloc/op   delta
Ingest-4    1.50kB ± 0%    1.31kB ± 0%  -12.84%  (p=0.008 n=5+5)

name      old allocs/op  new allocs/op  delta
Ingest-4      29.0 ± 0%      23.0 ± 0%  -20.69%  (p=0.000 n=6+5)
```
<!-- Type of change
Please label this PR with one of the following labels, depending on the scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
-->

## Why is it important?

It improves performance of the statsd module.

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] ~~I have made corresponding changes to the documentation~~
- [ ] ~~I have made corresponding change to the default configuration files~~
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] ~~I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.~~